plot-ai/backend/plot-backend/pyproject.toml
[project]
name = "plot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "django>=4.0,<5.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pydub>=0.25.1",
    "requests>=2.32.5",
    "openai-whisper>=20231117",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
